Update README; fix reqs
This adds instructions on setting up an environment for hacking on
Captain as well as how to run the tests.

It tweaks local.py-dist so that there are fewer things you have
to do to get going.

It tweaks requirements/dev.txt so that it includes all the other
requirements so there's only one step to install all requirements
required.