Add capability to log the output of commands executed by shove.
Captain now sends two extra pieces of info to shove: a log_key, which
identifies the log that the command's output should be written to, and
a logging_queue, which identifies the queue on which captain will listen
for logging events.

When shove finishes executing a command, it will send a message over the
queue that captain specified containing the output of the executed 
command. Captain, listening via a management command 
`monitor_log_events`, receives the message and writes it to a file in
the media directory.